refactor: 패키지 이동, 불필요한 import 삭제
- CarTest, GameTest를 test 내 domain 패키지로 이동
