+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,8 +1,7 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.Car;
 import racingcar.dto.CarInfo;
 
 import java.util.List;
